Change SQL

+++ conf/evolutions/default/1.sql
@@ -5,7 +5,7 @@ CREATE SEQUENCE users_id_seq START WITH 30000;
 CREATE SEQUENCE questions_id_seq START WITH 20000;
 
 CREATE TABLE users (
-    id long NOT NULL DEFAULT nextval('users_id_seq'),
+    id integer NOT NULL DEFAULT nextval('users_id_seq'),
     registration_date date,
     PRIMARY KEY (id)
 );
@@ -15,7 +15,7 @@ CREATE TABLE questions (
     title varchar(255),
     choice_A varchar(255),
     choice_B varchar(255),
-    author_id long,
+    author_id integer,
     total_A integer,
     total_B integer,
     creation_date date,
@@ -23,8 +23,8 @@ CREATE TABLE questions (
 );
 
 CREATE TABLE answers (
-    question_id long NOT NULL,
-    user_id long NOT NULL,
+    question_id integer NOT NULL,
+    user_id integer NOT NULL,
     choice char,
     creation_date date,
     PRIMARY KEY (question_id,user_id)
